[network_info_plus] adapt the Android plugin
