+++ packages/network_info_plus/android/build.gradle
@@ -1,4 +1,4 @@
-group 'io.flutter.plugins.connectivity'
+group 'dev.fluttercommunity.plus.network_info'
 version '1.0-SNAPSHOT'
 
 buildscript {
