+++ packages/network_info_plus/android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'connectivity'
+rootProject.name = 'network_info'
