--- packages/network_info_plus/android/src/main/java/dev/fluttercommunity/plus/network_info/ConnectivityBroadcastReceiver.java
@@ -1,50 +1,0 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-package dev.fluttercommunity.plus.connectivity;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.net.ConnectivityManager;
-import io.flutter.plugin.common.EventChannel;
-
-/**
- * The ConnectivityBroadcastReceiver receives the connectivity updates and send them to the UIThread
- * through an {@link EventChannel.EventSink}
- *
- * <p>Use {@link
- * io.flutter.plugin.common.EventChannel#setStreamHandler(io.flutter.plugin.common.EventChannel.StreamHandler)}
- * to set up the receiver.
- */
-class ConnectivityBroadcastReceiver extends BroadcastReceiver
-    implements EventChannel.StreamHandler {
-  private Context context;
-  private Connectivity connectivity;
-  private EventChannel.EventSink events;
-
-  ConnectivityBroadcastReceiver(Context context, Connectivity connectivity) {
-    this.context = context;
-    this.connectivity = connectivity;
-  }
-
-  @Override
-  public void onListen(Object arguments, EventChannel.EventSink events) {
-    this.events = events;
-    context.registerReceiver(this, new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION));
-  }
-
-  @Override
-  public void onCancel(Object arguments) {
-    context.unregisterReceiver(this);
-  }
-
-  @Override
-  public void onReceive(Context context, Intent intent) {
-    if (events != null) {
-      events.success(connectivity.getNetworkType());
-    }
-  }
-}
